Логика работы с БД перенесена в класс DataWorker, дополнена логика работы класса ValidationData. Поле Weight теперь имеет тип double и диапозон для ввода 0.1-70.

+++ Delivery Winform/Model/ResultModel/Abstractions/IResult.cs
@@ -10,7 +10,7 @@ namespace Delivery_Winform.Model.ResultModel.Abstractions
     {
         public int Id { get; set; }
         public int Id_OrderResult { get; set; }
-        public int Weight_OrderResult { get; set; }
+        public double Weight_OrderResult { get; set; }
         public int CityDistrict_OrderResult { get; set; }
         public DateTime DeliveryDateTime_OrderResult { get; set; }
     }
